5.8 Added Web

+++ app/src/main/java/com/example/siubo/wheretoyov50/MyItem.java
@@ -12,7 +12,7 @@ public class MyItem implements ClusterItem {
     private String mTitle;
     private String mSnippet;
     private String mComment;
-    private int is_private;
+    private long is_private;
 
     public MyItem(LatLng latlng) {
         mPosition = latlng;
@@ -26,7 +26,7 @@ public class MyItem implements ClusterItem {
         mSnippet = "";
     }
 
-    public MyItem(double lat, double lng, String hour, String lastseen, int is_private) {
+    public MyItem(double lat, double lng, String hour, String lastseen, long is_private) {
         mPosition = new LatLng(lat, lng);
         mTitle = hour;
         mSnippet = lastseen;
@@ -61,5 +61,5 @@ public class MyItem implements ClusterItem {
         return mSnippet;
     }
 
-    public int getIs_private() { return is_private; }
+    public long getIs_private() { return is_private; }
 }
